rename: 컴포넌트 이름 변경된 것 import도 수정
