+++ FE/src/components/common/index.ts
@@ -8,4 +8,4 @@ export { default as QuestionSelectionBox } from './QuestionSelectionBox/Question
 export { default as WorkbookCard } from './WorkbookCard/WorkbookCard';
 export { default as ResponsiveMenu } from './ResponsiveMenu/ResponsiveMenu';
 export { default as Mirror } from './Mirror/Mirror';
-export { default as InterviewSetGeneratorModal } from '@common/QuestionSelectionBox/WorkbookGeneratorModal/WorkbookGeneratorModal';
+export { default as WorkbookGeneratorModal } from '@common/QuestionSelectionBox/WorkbookGeneratorModal/WorkbookGeneratorModal';
